[example] improve(config): moved config object to separate file
